fix: react sdk jsx runtime
