+++ package.json
@@ -68,13 +68,13 @@
         "backend-test-sequential": "pnpm -F backend test-sequential",
         "frontend-dev": "pnpm -F frontend dev",
         "frontend-build": "pnpm -F frontend build",
-        "frontend-build-lib": "pnpm -F frontend build-lib",
+        "frontend-build-sdk": "pnpm -F frontend build-sdk",
         "frontend-lint": "pnpm -F frontend lint",
         "frontend-format": "pnpm -F frontend format",
         "frontend-fix-lint": "pnpm -F frontend fix-lint",
         "frontend-fix-format": "pnpm -F frontend fix-format",
         "frontend-test": "pnpm -F frontend test",
-        "sdk-build": "pnpm -F frontend build-lib && pnpm -F sdk build",
+        "sdk-build": "pnpm clean-build-cache && pnpm common-build && pnpm frontend-build-sdk",
         "e2e-open": "pnpm -F e2e cypress:open",
         "e2e-run": "pnpm -F e2e cypress:run",
         "e2e-lint": "pnpm -F e2e lint",
