+++ packages/backend/tsconfig.json
@@ -71,13 +71,5 @@
     "include": [
         "src/**/*.ts",
         "src/**/*.json"
-    ],
-    "references": [
-        {
-            "path": "../common"
-        },
-        {
-            "path": "../warehouses"
-        }
     ]
 }
