+++ packages/e2e/tsconfig.json
@@ -19,10 +19,5 @@
     },
     "include": [
         "**/*.ts"
-    ],
-    "references": [
-        {
-            "path": "../common"
-        }
     ]
 }
