+++ packages/frontend/vite.config.sdk.ts
@@ -1,0 +1,43 @@
+import { resolve } from 'path';
+import dts from 'vite-plugin-dts';
+import svgrPlugin from 'vite-plugin-svgr';
+import { defineConfig } from 'vitest/config';
+
+process.env.NODE_ENV = 'production';
+
+export default defineConfig({
+    publicDir: false,
+    define: {
+        __APP_VERSION__: JSON.stringify(process.env.npm_package_version),
+    },
+    plugins: [svgrPlugin(), dts({ rollupTypes: true })],
+    mode: 'production',
+    build: {
+        outDir: 'dist',
+        emptyOutDir: true,
+        target: 'es2020',
+        // minify: 'esbuild',
+        sourcemap: true,
+        lib: {
+            entry: resolve(__dirname, 'src', 'sdk', 'index.tsx'),
+            name: 'LightdashSDK',
+            formats: ['es', 'cjs'],
+            fileName: (ext) => `sdk.${ext}.js`,
+            cssFileName: 'sdk',
+        },
+        rollupOptions: {
+            external: [
+                'react',
+                'react-dom',
+                'react/jsx-runtime',
+                'react/jsx-dev-runtime',
+            ],
+            output: {
+                globals: {
+                    react: 'React',
+                    'react-dom': 'ReactDOM',
+                },
+            },
+        },
+    },
+});
