+++ packages/frontend/vite.config.ts
@@ -1,15 +1,12 @@
 import { sentryVitePlugin } from '@sentry/vite-plugin';
 import reactPlugin from '@vitejs/plugin-react';
 import { compression } from 'vite-plugin-compression2';
-import dts from 'vite-plugin-dts';
 import monacoEditorPlugin from 'vite-plugin-monaco-editor';
 import svgrPlugin from 'vite-plugin-svgr';
 import { defineConfig } from 'vitest/config';
 
-const isLib = process.env.VITE_LIB === 'true';
-
 export default defineConfig({
-    publicDir: isLib ? false : 'public',
+    publicDir: 'public',
     define: {
         __APP_VERSION__: JSON.stringify(process.env.npm_package_version),
     },
@@ -20,18 +17,10 @@ export default defineConfig({
             include: [/\.(js)$/, /\.(css)$/],
             filename: '[path][base].gzip',
         }),
-        ...(isLib
-            ? [
-                  dts({
-                      rollupTypes: true,
-                  }),
-              ]
-            : [
-                  monacoEditorPlugin({
-                      forceBuildCDN: true,
-                      languageWorkers: ['json'],
-                  }),
-              ]),
+        monacoEditorPlugin({
+            forceBuildCDN: true,
+            languageWorkers: ['json'],
+        }),
         sentryVitePlugin({
             org: 'lightdash',
             project: 'lightdash-frontend',
@@ -53,81 +42,61 @@ export default defineConfig({
         exclude: ['@lightdash/common'],
     },
     build: {
-        outDir: isLib ? 'dist' : 'build',
+        outDir: 'build',
+        emptyOutDir: false,
         target: 'es2020',
         minify: true,
         sourcemap: true,
-        ...(isLib
-            ? {
-                  lib: {
-                      entry: './src/sdk/index.tsx',
-                      name: '@lightdash/frontend',
-                      formats: ['es', 'cjs'],
-                      fileName: 'frontend',
-                  },
-                  rollupOptions: {
-                      external: [
-                          'react/jsx-runtime',
-                          'react-dom/jsx-runtime',
-                          'react/jsx-dev-runtime',
-                          'react-dom/jsx-dev-runtime',
-                          'react',
-                          'react-dom',
-                      ],
-                  },
-              }
-            : {
-                  rollupOptions: {
-                      output: {
-                          manualChunks: {
-                              react: [
-                                  'react',
-                                  'react-dom',
-                                  'react-router',
-                                  'react-hook-form',
-                                  'react-use',
-                                  // TODO: removed because of PNPM
-                                  // 'react-draggable',
-                                  '@hello-pangea/dnd',
-                                  '@tanstack/react-query',
-                                  '@tanstack/react-table',
-                                  '@tanstack/react-virtual',
-                              ],
-                              echarts: ['echarts'],
-                              vega: ['vega', 'vega-lite'],
-                              ace: ['ace-builds', 'react-ace/lib'],
-                              modules: [
-                                  // TODO: removed because of PNPM
-                                  // 'ajv',
-                                  // 'ajv-formats',
-                                  // 'liquidjs',
-                                  // 'pegjs',
-                                  'jspdf',
-                                  'lodash',
-                                  'colorjs.io',
-                                  'zod',
-                              ],
-                              thirdparty: [
-                                  '@sentry/react',
-                                  'rudder-sdk-js',
-                                  'posthog-js',
-                              ],
-                              uiw: [
-                                  '@uiw/react-markdown-preview',
-                                  '@uiw/react-md-editor',
-                              ],
-                              mantine: [
-                                  '@mantine/core',
-                                  '@mantine/dates',
-                                  '@mantine/form',
-                                  '@mantine/hooks',
-                                  '@mantine/notifications',
-                                  '@mantine/prism',
-                              ],
-                          },
-                      },
-                  },
-              }),
+
+        rollupOptions: {
+            output: {
+                manualChunks: {
+                    react: [
+                        'react',
+                        'react-dom',
+                        'react-router',
+                        'react-hook-form',
+                        'react-use',
+                        // TODO: removed because of PNPM
+                        // 'react-draggable',
+                        '@hello-pangea/dnd',
+                        '@tanstack/react-query',
+                        '@tanstack/react-table',
+                        '@tanstack/react-virtual',
+                    ],
+                    echarts: ['echarts'],
+                    ace: ['ace-builds', 'react-ace/lib'],
+                    modules: [
+                        // TODO: removed because of PNPM
+                        // 'ajv',
+                        // 'ajv-formats',
+                        // 'liquidjs',
+                        // 'pegjs',
+                        'jspdf',
+                        'lodash',
+                        'colorjs.io',
+                        'zod',
+                    ],
+                    thirdparty: [
+                        '@sentry/react',
+                        'rudder-sdk-js',
+                        'posthog-js',
+                    ],
+                    uiw: [
+                        '@uiw/react-markdown-preview',
+                        '@uiw/react-md-editor',
+                    ],
+                    mantine: [
+                        '@mantine/core',
+                        '@mantine/dates',
+                        '@mantine/form',
+                        '@mantine/hooks',
+                        '@mantine/notifications',
+                        '@mantine/prism',
+                    ],
+                },
+            },
+        },
     },
     test: {
         globals: true,
