+++ packages/sdk-test-app/tsconfig.json
@@ -19,9 +19,4 @@
     "include": [
         "src"
     ],
-    "references": [
-        {
-            "path": "../sdk"
-        }
-    ]
 }
