--- packages/sdk/tsconfig.json
@@ -1,33 +1,0 @@
-{
-  "extends": "./../../tsconfig.json",
-  "compilerOptions": {
-    "composite": true,
-    "target": "ES2020",
-    "lib": [
-      "ES2020",
-      "DOM",
-      "DOM.Iterable"
-    ],
-    "skipLibCheck": true,
-    /* Bundler mode */
-    "module": "ESNext",
-    "moduleResolution": "node",
-    "isolatedModules": true,
-    "moduleDetection": "force",
-    // "noEmit": true,
-    "jsx": "react-jsx",
-    /* Linting */
-    "strict": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noFallthroughCasesInSwitch": true,
-  },
-  "include": [
-    "src"
-  ],
-  "references": [
-    {
-      "path": "../frontend"
-    }
-  ]
-}
